src/03_DEPLOYMENT/FASTAPI/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
pyarrow
mlflow
scikit-learn
catboost
python-multipart
requests
boto3
python-dotenv